factor validation into gss
