+++ general-superstaq/general_superstaq/__init__.py
@@ -9,7 +9,7 @@ from general_superstaq.superstaq_exceptions import (
 )
 from general_superstaq.typing import MaxSharpeJson, MinVolJson, QuboModel, TSPJson, WareHouseJson
 
-from . import qubo, serialization, superstaq_client, superstaq_exceptions, user_config
+from . import qubo, serialization, superstaq_client, superstaq_exceptions, user_config, validation
 
 __all__ = [
     "__version__",
@@ -33,4 +33,5 @@ __all__ = [
     "superstaq_exceptions",
     "typing",
     "user_config",
+    "validation",
 ]
